Clean up some dependency issues
