+++ netlogo-gui/src/main/app/AppTabManager.scala
@@ -589,7 +589,7 @@ class AppTabManager(val appTabsPanel:          Tabs,
   // Prints list of tabs in a JTabbedPane
   def __printTabsOfTabsPanel(pane: JTabbedPane): Unit = {
     for (n <- 0 until pane.getTabCount) {
-      App.__printSwingObject(pane.getComponentAt(n), "")
+      __printSwingObject(pane.getComponentAt(n), "")
     }
   }
 
@@ -694,7 +694,7 @@ class AppTabManager(val appTabsPanel:          Tabs,
 
   // For App MenuBar - Prints Menus
   def __printAppMenuBar(): Unit = {
-    App.__printSwingObject(getAppMenuBar, "menu bar")
+    __printSwingObject(getAppMenuBar, "menu bar")
     for (i <- 0 until getAppMenuBar.getMenuCount) {
       val  item = getAppMenuBar.getMenu(i)
       if (item != null) {
@@ -745,6 +745,29 @@ class AppTabManager(val appTabsPanel:          Tabs,
       if (item.isInstanceOf[javax.swing.JMenu]) {
         __printMenuAccelerators(item.asInstanceOf[javax.swing.JMenu])
       }
+    }
+  }
+
+  def __printNonNullSwingObject(obj: java.awt.Component, description: String): Unit = {
+    val pattern = """(^[^\[]*)\[(.*$)""".r
+    val pattern(name, _) = obj.toString
+    val shortName = name.split("\\.").last
+    println(description + " " + System.identityHashCode(obj) +
+    ", " + shortName)
+  }
+
+  def __printSwingObject(obj: java.awt.Component, description: String): Unit = {
+    val some = Option(obj) // Because Option(null) = None 11/2020 AAB
+    some match {
+      case None           => println(description + " <null>")
+      case Some(theValue) =>  __printNonNullSwingObject(obj, description)
+    }
+  }
+
+  def __printOptionSwingObject(obj: Option[java.awt.Component], description: String): Unit = {
+    obj match {
+      case None            => println(description + " None")
+      case Some(theObject) =>  __printNonNullSwingObject(theObject, description)
     }
   }
